feat: change temp to double

+++ Structural/Adapter/Service/OldThermostat.java
@@ -7,7 +7,7 @@ public class OldThermostat {
         this.tempCelsius = temp; 
     }
 
-    public double getTempCelsius() {
+    public float getTempCelsius() {
         return tempCelsius; 
     }
 }
